nouvelle migration + suppression themes/fichiers inutiles

+++ wp-content/plugins/blocksy-companion/framework/dashboard.php
@@ -51,6 +51,12 @@ class Dashboard {
 
 		if (function_exists('blc_fs')) {
 			blc_fs()->add_filter('hide_plan_change', '__return_true');
+			blc_fs()->add_filter(
+				'plugin_icon',
+				function ($url) {
+					return BLOCKSY_PATH . '/static/img/logo.jpg';
+				}
+			);
 
 			blc_fs()->add_filter(
 				'permission_diagnostic_default',
@@ -195,6 +201,10 @@ class Dashboard {
 				return;
 			}
 
+			if (is_network_admin()) {
+				return;
+			}
+
 			if (intval(get_option('blc_activation_redirect', false)) === wp_get_current_user()->ID) {
 				delete_option('blc_activation_redirect');
 				exit(wp_redirect(admin_url('admin.php?page=ct-dashboard')));
@@ -295,6 +305,14 @@ class Dashboard {
 
 	public function is_dashboard_page() {
 		global $pagenow;
+
+		if (is_network_admin()) {
+			$is_ct_settings =
+				// 'themes.php' === $pagenow &&
+				isset( $_GET['page'] ) && 'blocksy-companion' === $_GET['page'];
+
+			return $is_ct_settings;
+		}
 
 		$is_ct_settings =
 			// 'themes.php' === $pagenow &&
